+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [18.0.0] - 2025-x-x
+### Added 
+* icon to ProviderSelection object in authorization flow requests
+
+### Changed
+* ⚠️ Breaking: replaced the withProviderSelection() flag on authorization flow request builder with a providerSelection(...) utility
+that helps to set the provider icon object if needed.
+
 ## [17.5.1] - 2025-10-31
 ### Fixed
 * Update Sonatype Central badge url in order to show latest version